requirements.txt
gradio>=4.0.0
numpy>=1.21.0
pillow>=8.0.0
tensorflow>=2.8.0
matplotlib>=3.5.0
